refactor: move client from utils
